Updates new-address-modal verifyAddress method
The value of this.address can also be null (occurs when cancelling the modal without any input), hence the check for address validity was previously still occurring when the modal was cancelled. This change ensures that we trap all falsey values for this.address, ensuring that we only check validity if a non-empty string is input.