Adds a startup error message to ensure that the connected daemon version is at least v0.18.0.6
With the release of core v0.18.x functionality such as the restart of the daemon to encrypt a created/restored wallet is no longer necessary. Hence a node needs to make use of a 0.18 version minimum.
Another change in 0.18.0.6 is that smsgmessages cannot be recieved from nodes running a core version less than 0.18.0.6. With the alpha mainnet release of the marketplace running on at least core v0.18.0.6 it is imperative, for the marketplace, to have the local node running this version.
While the desktop client will be released with at least this version bundled in, this change ensures that attempts to start this version with a running node on a lower version does not break the application.