Fixes wallet unlock modal resulting in multiple unlock emissions.
Specifically occurs due to the wallet being unlocked already (for staking only) when needing to unlock for sending functionality. In such cases, the state observer receieves the current status first, and the a 2nd status update when the wallet is unlocked, but in both cases, the wallet is unlocked and so regular processing occurs and an unlock event is emitted. Resulting in double functionality handling.

Secondly, when a wallet is unlocked for staking only and an unlock request for sending is made, a timeout is started in order to unlock the wallet for staking only after the timeout period. This is problematic, since the timeout is not stopped in any way once the wallet status is manually changed (via the user locking the wallet earlier than the timeout period), and thus causes the unlock for staking only to request to still be executed but at a user unpredictable time. This is problematic if the user manually locks and then unlocks the wallet before the timeout period expires: the wallet is unlocked for sending, and then suddenly the wallet is unlocked for staking only.

Both cases should now be resolved:
- an emit event is only emitted if the specific unlock status is differnt to the initial unlock status,
- if the initial status is unlocked for staking, then listen for an encryption status update on the wallet which where the wallet becomes locked, and only then attempt to set the wallet back to being unlocked for staking only.

This 2nd fix also fixes the case where the timeout would occur regardless of the wallet selected: if a wallet is unlocked for staking only, and then unlocked as a result of a user's send request, and then the user switches wallets, the timeout would occur regardless, and siimply fail because the unlock request would fail against the new selected wallet (different password, or the wallet is not encrypted). This change fixes this because the wallet status listener is completed for the current wallet when a wallet switch occurs.
HOWEVER: this does not solve a case where a wallet is unlocked for staking only, then unlocked by a user requested action, and before the wallet unlock for send timeout occurs, a wallet switch occurs to a different wallet and then back to the starting wallet. In such a case, the wallet will be unlocked and displayed as such, but once the timeout occurs the wallet will not change back to being unlocked for staking only.